requirements.txt
flask==3.0.0
flask-cors==4.0.0
ultralytics==8.1.0
opencv-python-headless==4.8.1.78
pillow==10.1.0
python-dotenv==1.0.0
torch==2.1.1
torchvision==0.16.1
numpy==1.24.3
gunicorn==21.2.0
